+++ dataset/GitHub_Java/Graylog2.graylog2-server/graylog2-server/src/main/java/org/graylog/plugins/sidecar/rest/resources/ConfigurationResource.java
@@ -16,6 +16,7 @@
  */
 package org.graylog.plugins.sidecar.rest.resources;
 
+import com.codahale.metrics.annotation.Timed;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.hash.Hashing;
 import io.swagger.annotations.Api;
@@ -25,28 +26,37 @@ import org.apache.shiro.authz.annotation.RequiresAuthentication;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.graylog.plugins.sidecar.audit.SidecarAuditEventTypes;
 import org.graylog.plugins.sidecar.permissions.SidecarRestPermissions;
+import org.graylog.plugins.sidecar.rest.models.CollectorUpload;
 import org.graylog.plugins.sidecar.rest.models.Configuration;
 import org.graylog.plugins.sidecar.rest.models.ConfigurationSummary;
 import org.graylog.plugins.sidecar.rest.models.Sidecar;
+import org.graylog.plugins.sidecar.rest.requests.ConfigurationAssignment;
 import org.graylog.plugins.sidecar.rest.requests.ConfigurationPreviewRequest;
+import org.graylog.plugins.sidecar.rest.responses.CollectorUploadListResponse;
 import org.graylog.plugins.sidecar.rest.responses.ConfigurationListResponse;
 import org.graylog.plugins.sidecar.rest.responses.ConfigurationPreviewRenderResponse;
-import org.graylog.plugins.sidecar.rest.responses.ValidationResponse;
+import org.graylog.plugins.sidecar.rest.responses.ConfigurationSidecarsResponse;
 import org.graylog.plugins.sidecar.services.ConfigurationService;
 import org.graylog.plugins.sidecar.services.EtagService;
+import org.graylog.plugins.sidecar.services.ImportService;
 import org.graylog.plugins.sidecar.services.SidecarService;
+import org.graylog.plugins.sidecar.template.RenderTemplateException;
 import org.graylog2.audit.jersey.AuditEvent;
 import org.graylog2.audit.jersey.NoAuditEvent;
 import org.graylog2.database.PaginatedList;
 import org.graylog2.plugin.rest.PluginRestResource;
+import org.graylog2.plugin.rest.ValidationResult;
 import org.graylog2.search.SearchQuery;
 import org.graylog2.search.SearchQueryField;
 import org.graylog2.search.SearchQueryParser;
 import org.graylog2.shared.rest.resources.RestResource;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import javax.inject.Inject;
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
+import javax.ws.rs.BadRequestException;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.DefaultValue;
@@ -64,36 +74,49 @@ import javax.ws.rs.core.EntityTag;
 import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
+
+import static com.google.common.base.MoreObjects.firstNonNull;
 
 @Api(value = "Sidecar/Configurations", description = "Manage/Render collector configurations")
 @Path("/sidecar/configurations")
 @Consumes(MediaType.APPLICATION_JSON)
 @Produces(MediaType.APPLICATION_JSON)
+@RequiresAuthentication
 public class ConfigurationResource extends RestResource implements PluginRestResource {
+    private static final Logger LOG = LoggerFactory.getLogger(ConfigurationResource.class);
+
+    // a file is created by the Sidecar based on the configuration name so we basically check for invalid paths here
+    private static final Pattern VALID_NAME_PATTERN = Pattern.compile("^[^;*?\"<>|&]+$");
+
     private final ConfigurationService configurationService;
     private final SidecarService sidecarService;
     private final EtagService etagService;
+    private final ImportService importService;
     private final SearchQueryParser searchQueryParser;
     private static final ImmutableMap<String, SearchQueryField> SEARCH_FIELD_MAPPING = ImmutableMap.<String, SearchQueryField>builder()
             .put("id", SearchQueryField.create(Configuration.FIELD_ID))
-            .put("backend_id", SearchQueryField.create(Configuration.FIELD_COLLECTOR_ID))
+            .put("collector_id", SearchQueryField.create(Configuration.FIELD_COLLECTOR_ID))
             .put("name", SearchQueryField.create(Configuration.FIELD_NAME))
             .build();
 
     @Inject
     public ConfigurationResource(ConfigurationService configurationService,
                                  SidecarService sidecarService,
-                                 EtagService etagService) {
+                                 EtagService etagService,
+                                 ImportService importService) {
         this.configurationService = configurationService;
         this.sidecarService = sidecarService;
         this.etagService = etagService;
+        this.importService = importService;
         this.searchQueryParser = new SearchQueryParser(Configuration.FIELD_NAME, SEARCH_FIELD_MAPPING);;
     }
 
     @GET
-    @RequiresAuthentication
     @RequiresPermissions(SidecarRestPermissions.CONFIGURATIONS_READ)
     @Produces(MediaType.APPLICATION_JSON)
     @ApiOperation(value = "List all configurations")
@@ -103,7 +126,7 @@ public class ConfigurationResource extends RestResource implements PluginRestRes
                                                         @ApiParam(name = "sort",
                                                                          value = "The field to sort the result on",
                                                                          required = true,
-                                                                         allowableValues = "name,id,backend_id")
+                                                                         allowableValues = "name,id,collector_id")
                                                                      @DefaultValue(Configuration.FIELD_NAME) @QueryParam("sort") String sort,
                                                         @ApiParam(name = "order", value = "The sort direction", allowableValues = "asc, desc")
                                                                      @DefaultValue("asc") @QueryParam("order") String order) {
@@ -118,41 +141,72 @@ public class ConfigurationResource extends RestResource implements PluginRestRes
     }
 
     @GET
+    @Path("/uploads")
+    @RequiresPermissions(SidecarRestPermissions.CONFIGURATIONS_READ)
+    @Produces(MediaType.APPLICATION_JSON)
+    @ApiOperation(value = "List all uploaded configurations")
+    public CollectorUploadListResponse listImports(@ApiParam(name = "page") @QueryParam("page") @DefaultValue("1") int page) {
+        // sort by creation date, latest on top of the list
+        final PaginatedList<CollectorUpload> uploads = this.importService.findPaginated(page, 10, "created", "desc");
+        final long total = this.importService.count();
+        final List<CollectorUpload> result = new ArrayList<>(uploads);
+
+        return CollectorUploadListResponse.create(uploads.pagination(), total, result);
+    }
+
+    @GET
     @Path("/{id}")
-    @RequiresAuthentication
     @RequiresPermissions(SidecarRestPermissions.CONFIGURATIONS_READ)
     @Produces(MediaType.APPLICATION_JSON)
     @ApiOperation(value = "Show configuration details")
     public Configuration getConfigurations(@ApiParam(name = "id", required = true)
-                                                    @PathParam("id") String id) {
-        return this.configurationService.find(id);
+                                           @PathParam("id") String id) {
+        final Configuration configuration = this.configurationService.find(id);
+        if (configuration == null) {
+            throw new NotFoundException("Could not find Configuration <" + id + ">.");
+        }
+        return configuration;
     }
 
     @GET
+    @Path("/{id}/sidecars")
+    @RequiresPermissions({SidecarRestPermissions.CONFIGURATIONS_READ, SidecarRestPermissions.SIDECARS_READ})
+    @Produces(MediaType.APPLICATION_JSON)
+    @ApiOperation(value = "Show sidecars using the given configuration")
+    public ConfigurationSidecarsResponse getConfigurationSidecars(@ApiParam(name = "id", required = true)
+                                                                      @PathParam("id") String id) {
+        final Configuration configuration = this.configurationService.find(id);
+        if (configuration == null) {
+            throw new NotFoundException("Could not find Configuration <" + id + ">.");
+        }
+        final List<String> sidecarsWithConfiguration = sidecarService.all().stream()
+                .filter(sidecar -> isConfigurationAssignedToSidecar(configuration.id(), sidecar))
+                .map(Sidecar::id)
+                .collect(Collectors.toList());
+        return ConfigurationSidecarsResponse.create(configuration.id(), sidecarsWithConfiguration);
+    }
+
+    @POST
     @Path("/validate")
-    @RequiresAuthentication
+    @NoAuditEvent("Validation only")
     @RequiresPermissions(SidecarRestPermissions.CONFIGURATIONS_READ)
     @Produces(MediaType.APPLICATION_JSON)
-    @ApiOperation(value = "Validates configuration name")
-    public ValidationResponse validateConfiguration(@ApiParam(name = "name", required = true) @QueryParam("name") String name) {
-        final Configuration configuration = this.configurationService.findByName(name);
-        if (configuration == null) {
-            return ValidationResponse.create(false, null);
-        }
-        return ValidationResponse.create(true, "Configuration with name \"" + name + "\" already exists");
+    @ApiOperation(value = "Validates configuration parameters")
+    public ValidationResult validateConfiguration(@Valid @ApiParam("configuration") Configuration toValidate) {
+        return validate(toValidate);
     }
 
     @GET
+    @Timed
     @Path("/render/{sidecarId}/{configurationId}")
     @Produces(MediaType.APPLICATION_JSON)
-    @RequiresAuthentication
     @RequiresPermissions(SidecarRestPermissions.CONFIGURATIONS_READ)
     @ApiOperation(value = "Render configuration template")
     public Response renderConfiguration(@Context HttpHeaders httpHeaders,
                                         @ApiParam(name = "sidecarId", required = true)
                                         @PathParam("sidecarId") String sidecarId,
                                         @ApiParam(name = "configurationId", required = true)
-                                        @PathParam("configurationId") String configurationId) {
+                                        @PathParam("configurationId") String configurationId) throws RenderTemplateException {
         String ifNoneMatch = httpHeaders.getHeaderString("If-None-Match");
         Boolean etagCached = false;
         Response.ResponseBuilder builder = Response.noContent();
@@ -202,67 +256,96 @@ public class ConfigurationResource extends RestResource implements PluginRestRes
     @POST
     @Path("/render/preview")
     @Produces(MediaType.APPLICATION_JSON)
-    @RequiresAuthentication
     @RequiresPermissions(SidecarRestPermissions.CONFIGURATIONS_READ)
     @ApiOperation(value = "Render preview of a configuration template")
     @NoAuditEvent("this is not changing any data")
     public ConfigurationPreviewRenderResponse renderConfiguration(@ApiParam(name = "JSON body", required = true)
                                                                   @Valid @NotNull ConfigurationPreviewRequest request) {
-        String preview = this.configurationService.renderPreview(request.template());
-        return ConfigurationPreviewRenderResponse.create(preview);
+        try {
+            String preview = this.configurationService.renderPreview(request.template());
+            return ConfigurationPreviewRenderResponse.create(preview);
+        } catch (RenderTemplateException e) {
+            throw new BadRequestException("Could not render template preview: " + e.getMessage());
+        }
     }
 
     @POST
-    @RequiresAuthentication
     @RequiresPermissions(SidecarRestPermissions.CONFIGURATIONS_CREATE)
     @Produces(MediaType.APPLICATION_JSON)
     @ApiOperation(value = "Create new configuration")
     @AuditEvent(type = SidecarAuditEventTypes.CONFIGURATION_CREATE)
-    public Configuration createConfiguration(@ApiParam(name = "JSON body", required = true)
+    public Response createConfiguration(@ApiParam(name = "JSON body", required = true)
                                              @Valid @NotNull Configuration request) {
-        Configuration configuration = configurationService.fromRequest(request);
-        return configurationService.save(configuration);
+        final Configuration configuration = configurationFromRequest(null, request);
+        final ValidationResult validationResult = validate(configuration);
+        if (validationResult.failed()) {
+            return Response.status(Response.Status.BAD_REQUEST).entity(validationResult).build();
+        }
+
+        return Response.ok().entity(configurationService.save(configuration)).build();
     }
 
     @POST
     @Path("/{id}/{name}")
-    @RequiresAuthentication
-    @RequiresPermissions(SidecarRestPermissions.CONFIGURATIONS_CREATE)
+    @RequiresPermissions({SidecarRestPermissions.CONFIGURATIONS_READ, SidecarRestPermissions.CONFIGURATIONS_CREATE})
     @ApiOperation(value = "Copy a configuration")
     @AuditEvent(type = SidecarAuditEventTypes.CONFIGURATION_CLONE)
     public Response copyConfiguration(@ApiParam(name = "id", required = true)
                                       @PathParam("id") String id,
                                       @PathParam("name") String name) throws NotFoundException {
         final Configuration configuration = configurationService.copyConfiguration(id, name);
+        final ValidationResult validationResult = validate(configuration);
+        if (validationResult.failed()) {
+            return Response.status(Response.Status.BAD_REQUEST).entity(validationResult).build();
+        }
         configurationService.save(configuration);
         return Response.accepted().build();
     }
 
     @PUT
     @Path("/{id}")
-    @RequiresAuthentication
     @RequiresPermissions(SidecarRestPermissions.CONFIGURATIONS_UPDATE)
     @Produces(MediaType.APPLICATION_JSON)
     @ApiOperation(value = "Update a configuration")
     @AuditEvent(type = SidecarAuditEventTypes.CONFIGURATION_UPDATE)
-    public Configuration updateConfiguration(@ApiParam(name = "id", required = true)
+    public Response updateConfiguration(@ApiParam(name = "id", required = true)
                                              @PathParam("id") String id,
                                              @ApiParam(name = "JSON body", required = true)
                                              @Valid @NotNull Configuration request) {
+        final Configuration previousConfiguration = configurationService.find(id);
+        if (previousConfiguration == null) {
+            throw new NotFoundException("Could not find Configuration <" + id + ">.");
+        }
+
+        // Only allow changing the associated collector ID if the configuration is not in use
+        if (!previousConfiguration.collectorId().equals(request.collectorId())) {
+            if (isConfigurationInUse(id)) {
+                throw new BadRequestException("Configuration still in use, cannot change collector type.");
+            }
+        }
+
+        final Configuration updatedConfiguration = configurationFromRequest(id, request);
+        final ValidationResult validationResult = validate(updatedConfiguration);
+        if (validationResult.failed()) {
+            return Response.status(Response.Status.BAD_REQUEST).entity(validationResult).build();
+        }
         etagService.invalidateAll();
-        Configuration configuration = configurationService.fromRequest(id, request);
-        return configurationService.save(configuration);
+
+        return Response.ok().entity(configurationService.save(updatedConfiguration)).build();
     }
 
     @DELETE
     @Path("/{id}")
-    @RequiresAuthentication
-    @RequiresPermissions(SidecarRestPermissions.CONFIGURATIONS_UPDATE)
+    @RequiresPermissions(SidecarRestPermissions.CONFIGURATIONS_DELETE)
     @Produces(MediaType.APPLICATION_JSON)
     @ApiOperation(value = "Delete a configuration")
     @AuditEvent(type = SidecarAuditEventTypes.CONFIGURATION_DELETE)
     public Response deleteConfiguration(@ApiParam(name = "id", required = true)
                                         @PathParam("id") String id) {
+        if (isConfigurationInUse(id)) {
+            throw new BadRequestException("Configuration still in use, cannot delete.");
+        }
+
         int deleted = configurationService.delete(id);
         if (deleted == 0) {
             return Response.notModified().build();
@@ -271,9 +354,69 @@ public class ConfigurationResource extends RestResource implements PluginRestRes
         return Response.accepted().build();
     }
 
+    private ValidationResult validate(Configuration toValidate) {
+        final Optional<Configuration> configurationOptional;
+        final Configuration configuration;
+        final ValidationResult validation = new ValidationResult();
+
+        configurationOptional = Optional.ofNullable(configurationService.findByName(toValidate.name()));
+        if (configurationOptional.isPresent()) {
+            configuration = configurationOptional.get();
+            if (!configuration.id().equals(toValidate.id())) {
+                // a configuration exists with a different id, so the name is already in use, fail validation
+                validation.addError("name", "Configuration \"" + toValidate.name() + "\" already exists");
+            }
+        }
+
+        if (toValidate.name().isEmpty()) {
+            validation.addError("name", "Configuration name cannot be empty.");
+        } else if (!VALID_NAME_PATTERN.matcher(toValidate.name()).matches()) {
+                validation.addError("name", "Configuration name can not include the following characters: ; * ? \" < > | &");
+        }
+
+        if (toValidate.collectorId().isEmpty()) {
+            validation.addError("collector_id", "Associated collector ID cannot be empty.");
+        }
+
+        if (toValidate.color().isEmpty()) {
+            validation.addError("color", "Collector color cannot be empty.");
+        }
+
+        if (toValidate.template().isEmpty()) {
+            validation.addError("template", "Collector template cannot be empty.");
+        }
+
+        try {
+            this.configurationService.renderPreview(toValidate.template());
+        } catch (RenderTemplateException e) {
+            validation.addError("template", "Template error: " + e.getMessage());
+        }
+
+        return validation;
+    }
+
+    private boolean isConfigurationInUse(String configurationId) {
+        return sidecarService.all().stream().anyMatch(sidecar -> isConfigurationAssignedToSidecar(configurationId, sidecar));
+    }
+
+    private boolean isConfigurationAssignedToSidecar(String configurationId, Sidecar sidecar) {
+        final List<ConfigurationAssignment> assignments = firstNonNull(sidecar.assignments(), new ArrayList<>());
+        return assignments.stream().anyMatch(assignment -> assignment.configurationId().equals(configurationId));
+    }
+
     private String configurationToEtag(Configuration configuration) {
         return Hashing.md5()
                 .hashInt(configuration.hashCode())  // avoid negative values
                 .toString();
+    }
+
+    private Configuration configurationFromRequest(String id, Configuration request) {
+        Configuration configuration;
+        if (id == null) {
+            configuration = configurationService.fromRequest(request);
+        } else {
+            configuration = configurationService.fromRequest(id, request);
+        }
+        return configuration;
     }
 }
